Big rework of the build process to allow building any of the following:   - A BlackICE only ROM from an existing CM7 ROM (no CM7 source required)   - A CM7 ROM (from source) and then a BlackICE ROM built on top of that   - A CM7 ROM only (from source)
tools/build.sh is the master build script again, but with some changes
  - The build parameters have been reworked and can come from a .ini file
    (see conf/build_example.ini) or the command line.
  - build_example.ini contains many comments for using the various parameters.
  - The directory ICEDroid/build_scripts has been added and contains
    * init.sh - processes all the .ini and command line parameters
    * build_blackice.sh - handles BlackICE specific build items (this is
      basically the old version of build.sh)
    * build_cm7.sh - handles CM7 specific build items

A few other scripts had to be modified to play nicely with the above changes.

bacon.sh and bacon.ini have been deleted (build_example.ini and build.sh do
everything they used to do).

The patch files in ICEDroid/src have been renamed to follow a stricter format
(see build_example.ini for more details) and I turned on eXecute permission
on them so it is possible to actually run them. I have tested the 2 existing
patch files and verified that it is now possible for patch files to be applied
during the build process (as long as they are named properly).

There is still some cosmetic cleanup that could be done, for example, the log
mechanism used by build_blackice.sh should probably be used by the rest of the
build process and should be initated either by init.sh or by build.sh *after*
ini.sh has run and all the variable are set. Or build.sh could turn on logging
right away and just put the log in whatever the current directory is.