Build process changes - mainly how patches are managed
1) '-dbox' can be used to set DROPBOX_DIR to a path (presumably inside your
public dropbox) to copy the result CM7 and/or BlackICE Kangs into. This is
useful if you are building something that you are going to post a link to
because it starts the upload to Dropbox as soon as possible (before pushing to
your phone) since that takes a "long" time.

The default is a NULL value and no files will be copied to your Dropbox.

2) '-patch' has been removed. Instead patches are loaded from the new file
ICEDroid/src/default_patches. See that file for more information on how to
define a new patch. The first field of a patch definition is a 'name' that
should be in all upper case, like TORCH. This will cause the build process to
look for the variable PATCH_TORCH and only apply the patch if this variable is
1 (the default case). To disable a defined patch set its variable to 0 in the
.ini file. See blackice_example.ini.

3) By default patches are only applied if the source you are building is also
being synced. In otherwords CM7 patches will only be applied if the build is
also going to do a 'repo sync' and BlackICE patches will only be applied if
the build is also going to do a 'git pull'. This helps eliminates errors that
occur when you try to apply a patch on the same code multiple times. (I've
only tested this for CM7 patches because we don't have any BlackICE patches
at the moment. If you do a 'repo sync' it discards any previous cherry-picks
so it will work fine to apply the patches after that).

If you want to override this you can use the '-fpatch' command line option.
The main use of this is if you manually did a 'repo sync' (which discarded
any previous patches) and then you decide to do a build, but don't want to
sync again.

4) '-check' has been replaced with '-prompt'. The new default when building
is to prompt the user with a 10 second countdown before building. This
(hopefully) gives them enough time to detect if the build options are not
correct and press CTRL-C. To skip the prompt and build immediately you can
use the '-prompt 0' option.

5) Sorry if I didn't provide enough documentation :)