Set version to v1.27b4 for nightly build, minor build changes
The build changes handle the creation of md5sums.txt much better.
Changing into the release directory before generating the sums
means the there is no path on the file name and so no extra
processing is needed to remove it now.

Minor changes to the format of changes.txt