Fixed an issue processing the density in autorun.txt
The check for a valid density value was previously failing if
there was extra whitespace around the number. Since density
is the first option in autorun.txt this means none of the
other options would get processed either. I modified the
check to be tolerant of white space and it is working
for me now. I also tested gpsconf, ril and braviaengine
all with extra white space and that works fine.