Added framework-res.4.* added an updated TORCH patch
- The latest CM7 changes modified the file that the previous TORCH
patch operated on. I had to create a new TORCH patch that uses diff
instead of cherry-picking the change. This works, but it may make
updating and repatching CM7 slightly more difficult until I find
a good way for the build process to deal with this.

- The framework-res.4 includes some changes from The Unknown Noob
for the progress bar and volume up/down bar. If you wanted to build
the old way for some reason you can rename framework-res.3.exclude
to framework-res.4.exclude in your working copy. I doubt we need
this and I imagine that soon the entire framework-res.3 will go away.