Changes related to patching
- Added ...TORCH1.git patch that is used *before* ...TORCH1.patch.
This does a 'git checkout -f' of the file about to be patched so
that we have a clean unpatched file to work with. Otherwise there
will be a build error.

- Fixed a bug that tested FORCE_PATCHING for the wrong value, which
prevented -fpatch from actually working and of course I needed it
to work to test my other patch changes. I guess it's only fair that
I run into a problem because of my own bug...

- version bumped to v1.27b7