Allow the build to directly replace SystemUI.apk instead trying to mod it.
Added support for a 'replace' method in the xxx.options files. This allows
a .apk to get directly replaced instead of processed. Of course this is
not ideal if that .apk changes a lot, but this will allow using The
Unknown Noobs new graphics right away until we find a better way to
do this.

NOTE: If you want to build the old way simply rename SystemUI.exclude
to SystemUI.1.exclude.