Reverted most of my previous change because it was wrong.
There was already a way to do what I wanted, but I didn't realize
it. The new SystemUI.apk is now in the /system/app directory of
BlackICE, which causes it to be copied directly into the build.
Then we only need a SystemUI.exclude file to keep the old
SystemUI.apk from being processed.

The SystemUI.apk that is now in /system/app is brand new from
The Unknown Noob.