Modified the file names When '-official 1' is used.
Previously I changed official builds to use OFFICIAL instead of a
timestamp, but that is not consistent with nightly builds. I changed
this to keep the normal timestamp, but prefix it with 'RC-' for
Release Candidate.

Minor format changes to the changes.txt file that is created as
part of a build.