The build process supports CM9 (ICS), but NOT ICS+BlackICE!
I used CM9 instead of ICS just because it was more convenient and
orthagonal with the CM7 the build process already supported.

The reason for this change was to make it convenient and possible
to build a Kang for CM7, CM9, BlackICE or CM7+BlackICE all from
the same build process.

This should not disturb any existing stuff except that a few
build variables and command line arguments got renamed to work
better with both CM7 and CM9.