Reworked refreshapps.sh, changed wget to use -nv (not verbose)
refreshapps.sh now pulls all matching .apk files from the phone into a
temp directory and then does a diff against the existing ones in
the download directory. The temp directory is left with only the
.apk files that are different and which should be uploaded to
buildpackages on the server.

Renamed the orginal refreshapps.sh to refreshapps_old.sh in case
it is still needed.